requirements.txt
appdirs==1.4.3
attrs==19.3.0
Automat==0.8.0
backports.functools-lru-cache==1.6.1
backports.os==0.1.1
certifi==2019.11.28
cffi==1.14.0
cfscrape==2.0.8
chardet==3.0.4
Click==7.0
colorama==0.4.3
coloredlogs==10.0
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
docopt==0.6.2
enum34==1.1.6
en_core_web_sm==2.2.0
Flask==1.1.1
flask-mongoengine==0.9.5
flask_restful==0.3.8
flask-bcrypt==0.7.1
flask-jwt-extended==3.24.1
fs==2.4.11
future==0.17.1
fuzzywuzzy==0.17.0
humanfriendly==4.18
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipaddress==1.0.23
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.5.0
MarkupSafe==1.1.1
monotonic==1.5
parsel==1.5.2
pigar==0.9.2
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pycryptodome==3.9.4
PyDispatcher==2.0.5
PyHamcrest==1.10.1
pymongo==3.10.1
pyOpenSSL==19.1.0
python-dateutil==2.8.0
pytz==2019.2
queuelib==1.5.0
requests==2.22.0
requests-cache==0.5.2
service-identity==18.1.0
six==1.14.0
soupsieve==1.9.5
spacy==2.2.3
tabulate==0.8.6
tenacity==5.1.1
tqdm==4.36.1
Twisted==19.10.0
typing==3.7.4.1
urllib3==1.25.8
verboselogs==1.7
w3lib==1.21.0
Werkzeug==1.0.0
yarg==0.1.9
zope.interface==4.7.1
